// src/huellas/dto/huella-response.dto.ts
export class HuellaResponseDto {

    id_huella_ciudadano: number;

    ciudadano_id: number;

    dedo_id: number;

    activo: boolean;

    detalle_motivo: string;

    fecha_registro: Date;

    fecha_modificacion: Date;

    organismo_id: number;

    usuario_id: number;
}